backend/requirements.txt
# Backend API dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Core AI/ML dependencies
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
lxml>=4.9.0
aiofiles>=23.2.1

# Development dependencies
black>=23.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
